Add "rename symbol at point" bound to F2 key
A function to automate renaming the same name "symbol" in the entire
function, so that I don't have to go to the top of the file and then
manually revert back to where I was.

I've bound this to the F2 key (because it means "rename" in a lot of
software).

I've moved the "dired-hide-details-mode" binding (that was on F2) to F8.