Symbol highlighting requires 2 occurances
Highlighting doesn't start until there are 2 occurances, or more.

I've also decreased the number of messages printed to the mini-buffer -
and put the number of occurances in the mode line instead.

Also fixed a subtle bug with the cursor moving to the first character of
a symbol when moving to a previous occurances and the cursor is not on
the first character of the symbol.